DE_end-to-end/requirements.txt
faker==28.1.0
apache-airflow==2.10.0
apache-airflow-providers-microsoft-azure==10.4.0
apache-airflow-providers-postgres==5.12.0
apache-airflow-providers-apache-kafka==1.6.0
apache-airflow-providers-apache-spark==4.10.0
kafka-python-ng # https://github.com/dpkp/kafka-python/issues/2412
pyspark==3.5.2
azure-storage-blob==12.22.0
python-dotenv==1.0.1
matplotlib==3.9.2
seaborn==0.13.2
python-dotenv==1.0.1
numpy==1.26.4
scipy==1.12.0
pyarrow==17.0.0
ua-parser==0.18.0
user-agents==2.2.0
requests==2.32.3
psycopg2==2.9.9